config/weaviate-local/llm-kungfu/requirements.txt
python-dotenv
anthropic
requests
google-generativeai
